backend_django/requirements.txt
asgiref==3.11.0
Django==5.2.8
django-environ==0.12.0
djangorestframework==3.16.1
PyJWT==2.10.1
sqlparse==0.5.4
tzdata==2025.2
